requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.11.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
distlib==0.3.6
filelock==3.12.2
Flask==2.1.2
future==0.18.3
gunicorn==20.1.0
idna==3.4
imgurpython==1.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.3.0
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
pexpect==4.8.0
platformdirs==3.7.0
ptyprocess==0.7.0
pycparser==2.21
pydub==0.25.1
pyOpenSSL==23.2.0
requests==2.31.0
soupsieve==2.4.1
tqdm==4.65.0
urllib3==2.0.4
virtualenv==20.23.1
SpeechRecognition
pydub
# Werkzeug==2.3.6
yarl==1.9.2
# zipp==3.16.2
